Add messages.pot update script to pre-commit. (#8900)
* Add messages.pot update script to pre-commit.

* Update .pre-commit-config.yaml web.py version


* Update .pre-commit-config.yaml to prevent multiple runs.

* Fix incorrect variable type assignment.

* Add logic to only update messages.pot if something has changed.

* Add typing to set compare

* Re-apply file type selection following merge

* Simplify/clarify function syntax and remove linenos from messages.pot

* Add comment re: updating dependencies

* Add verbose and forced options to messages extract script

* Update webpy version following rebase

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Exclude untracked files from messages.pot generation

* Modify get_untracked_files to work with multiple extensions

+++ openlibrary/i18n/__init__.py
@@ -1,8 +1,10 @@
 import os
 import shutil
 import sys
+import subprocess
 from collections.abc import Iterator
 from io import BytesIO
+from pathlib import Path
 
 import web
 
@@ -31,6 +33,23 @@ def success_color_fn(text: str) -> str:
 def warning_color_fn(text: str) -> str:
     """Styles the text for printing to console with warning color."""
     return '\033[93m' + text + '\033[0m'
+
+
+def get_untracked_files(dirs: list[str], extensions: tuple[str, str] | str) -> set:
+    """Returns a set of all currently untracked files with specified extension(s)."""
+    untracked_files = {
+        Path(line)
+        for dir in dirs
+        for line in subprocess.run(
+            ['git', 'ls-files', '--others', '--exclude-standard', dir],
+            stdout=subprocess.PIPE,
+            text=True,
+            check=True,
+        ).stdout.split('\n')
+        if line.endswith(extensions)
+    }
+
+    return untracked_files
 
 
 def _compile_translation(po, mo):
@@ -139,10 +158,15 @@ def extract_templetor(fileobj, keywords, comment_tags, options):
     return extract_python(f, keywords, comment_tags, options)
 
 
-def extract_messages(dirs: list[str]):
+def extract_messages(dirs: list[str], verbose: bool, forced: bool):
     catalog = Catalog(project='Open Library', copyright_holder='Internet Archive')
     METHODS = [("**.py", "python"), ("**.html", "openlibrary.i18n:extract_templetor")]
     COMMENT_TAGS = ["NOTE:"]
+
+    untracked_files = get_untracked_files(dirs, ('.py', '.html'))
+    template = read_po((Path(root) / 'messages.pot').open('rb'))
+    msg_set = {msg.id for msg in template if msg.id != ''}
+    new_set = set()
 
     for d in dirs:
         extracted = extract_from_dir(
@@ -151,19 +175,30 @@ def extract_messages(dirs: list[str]):
 
         counts: dict[str, int] = {}
         for filename, lineno, message, comments, context in extracted:
+            file_path = Path(d) / filename
+            if file_path in untracked_files:
+                continue
+            new_set.add(message)
             counts[filename] = counts.get(filename, 0) + 1
             catalog.add(message, None, [(filename, lineno)], auto_comments=comments)
 
-        for filename, count in counts.items():
-            path = filename if d == filename else os.path.join(d, filename)
-            print(f"{count}\t{path}", file=sys.stderr)
+        if verbose:
+            for filename, count in counts.items():
+                path = filename if d == filename else os.path.join(d, filename)
+                print(f"{count}\t{path}", file=sys.stderr)
 
-    path = os.path.join(root, 'messages.pot')
-    f = open(path, 'wb')
-    write_po(f, catalog)
-    f.close()
+    has_changed = msg_set != new_set
+    if has_changed or forced:
+        path = os.path.join(root, 'messages.pot')
+        f = open(path, 'wb')
+        write_po(f, catalog, include_lineno=False)
+        f.close()
 
-    print('wrote template to', path)
+        print('Updated strings written to', path)
+    else:
+        print(
+            'No modified strings discovered. To force a template update, re-run with --force-write.'
+        )
 
 
 def compile_translations(locales: list[str]):
